+++ changelog.md
@@ -115,6 +115,7 @@ Codion Change Log
 - SliderBuilder.ModelBuilder added.
 - ComboBoxBuilder.ModelBuilder added.
 - ItemComboBoxBuilder.BuilderFactory added.
+- Components.temporalField() parameter removed.
 ### is.codion.swing.framework.model
 - EntityComboBoxModel.Builder.EntityTypeBuilder and ConnectionProviderBuilder added.
 ### is.codion.swing.framework.ui